+++ cromwell-drs-localizer/src/test/scala/drs/localizer/downloaders/BulkAccessUrlDownloaderSpec.scala
@@ -1,11 +1,14 @@
 package drs.localizer.downloaders
 
-import cats.effect.IO
+import cats.effect.{ExitCode, IO}
 import cloud.nio.impl.drs.{AccessUrl, DrsResolverResponse}
 import common.assertion.CromwellTimeoutSpec
+import org.scalatest.prop.TableDrivenPropertyChecks._
 import drs.localizer.{ResolvedDrsUrl, URIType}
+
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
+
 import java.nio.file.Path
 
 class BulkAccessUrlDownloaderSpec extends AnyFlatSpec with CromwellTimeoutSpec with Matchers {
@@ -75,5 +78,37 @@ class BulkAccessUrlDownloaderSpec extends AnyFlatSpec with CromwellTimeoutSpec w
     val filepath: Path = downloader.generateJsonManifest(threeElements).unsafeRunSync()
     val expected = s"""getm --manifest ${filepath.toString}"""
     downloader.generateGetmCommand(filepath) shouldBe expected
+  }
+
+  {
+    val results = Table(
+      ("exitCode", "stderr", "download result"),
+      (0, "", DownloadSuccess),
+      // In `getm` version 0.0.4 checksum failures currently exit 0.
+      (0, "oh me oh my: AssertionError: Checksum failed!!!", ChecksumFailure),
+      // Unrecognized because of non-zero exit code without an HTTP status, despite what looks like a checksum failure.
+      (1, "oh me oh my: AssertionError: Checksum failed!!!", UnrecognizedRetryableDownloadFailure(ExitCode(1))),
+      // Unrecognized because of zero exit status with stderr that does not look like a checksum failure.
+      (0, "what the", UnrecognizedRetryableDownloadFailure(ExitCode(0))),
+      // Unrecognized because of non-zero exit code without an HTTP status.
+      (1, " foobar ", UnrecognizedRetryableDownloadFailure(ExitCode(1))),
+      // Unrecognized because of zero exit status with stderr that does not look like a checksum failure.
+      (0, """ERROR:getm.cli possibly some words "status_code": 503 words""", UnrecognizedRetryableDownloadFailure(ExitCode(0))),
+      // Recognized because of non-zero exit status and an HTTP status.
+      (1, """ERROR:getm.cli possibly some words "status_code": 503 words""", RecognizedRetryableDownloadFailure(ExitCode(1))),
+      // Recognized because of non-zero exit status and an HTTP status.
+      (1, """ERROR:getm.cli possibly some words "status_code": 408 more words""", RecognizedRetryableDownloadFailure(ExitCode(1))),
+      // Recognized and non-retryable because of non-zero exit status and 404 HTTP status.
+      (1, """ERROR:getm.cli possibly some words "status_code": 404 even more words""", FatalDownloadFailure(ExitCode(1))),
+      // Unrecognized because of zero exit status and 404 HTTP status.
+      (0, """ERROR:getm.cli possibly some words "status_code": 404 even more words""", UnrecognizedRetryableDownloadFailure(ExitCode(0))),
+    )
+    val bulkDownloader = BulkAccessUrlDownloader(null)
+
+    forAll(results) { (exitCode, stderr, expected) =>
+      it should s"produce $expected for exitCode $exitCode and stderr '$stderr'" in {
+        bulkDownloader.toDownloadResult(GetmResult(exitCode, stderr)) shouldBe expected
+      }
+    }
   }
 }
